# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ozone High School Payment System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: students
    description: Student management
  - name: payments
    description: Payment operations
  - name: admin
    description: Admin operations
  - name: cashier
    description: Cashier operations
  - name: dashboard
    description: Dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students (parent sees their children, admin/cashier sees all)
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"

  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments (parent sees own, cashier/admin sees all)
      parameters:
        - name: studentId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentWithStudent"

    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a Chapa payment (initiate checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment created with checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapaInitResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentWithStudent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payments/{id}/approve:
    patch:
      operationId: approvePayment
      tags: [cashier]
      summary: Cashier approves a pending manual payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentReviewInput"
      responses:
        "200":
          description: Payment approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentWithStudent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payments/{id}/reject:
    patch:
      operationId: rejectPayment
      tags: [cashier]
      summary: Cashier rejects a pending manual payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentReviewInput"
      responses:
        "200":
          description: Payment rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentWithStudent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payments/manual:
    post:
      operationId: submitManualPayment
      tags: [payments]
      summary: Submit a manual bank receipt upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualPaymentInput"
      responses:
        "201":
          description: Manual payment submitted for review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentWithStudent"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payments/chapa/verify/{txRef}:
    get:
      operationId: verifyChapaPayment
      tags: [payments]
      summary: Verify and record a Chapa payment after redirect
      parameters:
        - name: txRef
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment verified and recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentWithStudent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /admin/students:
    post:
      operationId: createStudent
      tags: [admin]
      summary: Create a new student (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/link-student:
    post:
      operationId: linkStudentToParent
      tags: [admin]
      summary: Link a student to a parent (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkStudentInput"
      responses:
        "200":
          description: Student linked to parent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [parent, cashier, admin]

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [parent, cashier, admin]
        createdAt:
          type: string

    Student:
      type: object
      required: [id, name, grade, section, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        grade:
          type: string
        section:
          type: string
        parentId:
          type: ["integer", "null"]
        parentName:
          type: ["string", "null"]
        createdAt:
          type: string

    StudentInput:
      type: object
      required: [name, grade, section]
      properties:
        name:
          type: string
        grade:
          type: string
        section:
          type: string
        parentId:
          type: ["integer", "null"]

    LinkStudentInput:
      type: object
      required: [studentId, parentId]
      properties:
        studentId:
          type: integer
        parentId:
          type: integer

    PaymentWithStudent:
      type: object
      required: [id, studentId, studentName, amount, method, status, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        studentGrade:
          type: string
        userId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [chapa, manual]
        status:
          type: string
          enum: [pending, approved, rejected]
        transactionRef:
          type: ["string", "null"]
        receiptImageUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string

    PaymentInput:
      type: object
      required: [studentId, amount]
      properties:
        studentId:
          type: integer
        amount:
          type: number

    ManualPaymentInput:
      type: object
      required: [studentId, amount, transactionRef]
      properties:
        studentId:
          type: integer
        amount:
          type: number
        transactionRef:
          type: string
        receiptImageUrl:
          type: ["string", "null"]

    ChapaInitResponse:
      type: object
      required: [checkoutUrl, txRef, paymentId]
      properties:
        checkoutUrl:
          type: string
        txRef:
          type: string
        paymentId:
          type: integer

    PaymentReviewInput:
      type: object
      properties:
        notes:
          type: ["string", "null"]

    DashboardStats:
      type: object
      required: [totalStudents, totalPayments, pendingPayments, approvedPayments, totalAmountCollected]
      properties:
        totalStudents:
          type: integer
        totalPayments:
          type: integer
        pendingPayments:
          type: integer
        approvedPayments:
          type: integer
        totalAmountCollected:
          type: number
        recentPayments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentWithStudent"