# lib/api-zod/src/generated/types/paymentReviewInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ozone High School Payment System API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentReviewInput {
  /** @nullable */
  notes?: string | null;
}