# lib/api-zod/src/generated/types/paymentWithStudent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ozone High School Payment System API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentWithStudentMethod } from "./paymentWithStudentMethod";
import type { PaymentWithStudentStatus } from "./paymentWithStudentStatus";

export interface PaymentWithStudent {
  id: number;
  studentId: number;
  studentName: string;
  studentGrade?: string;
  userId?: number;
  amount: number;
  method: PaymentWithStudentMethod;
  status: PaymentWithStudentStatus;
  /** @nullable */
  transactionRef?: string | null;
  /** @nullable */
  receiptImageUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
}