# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ozone High School Payment System API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  name: string;
  grade: string;
  section: string;
  /** @nullable */
  parentId?: number | null;
}